+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 # Development version
 
+- Parse errors in your document no longer trigger an LSP error when you request
+  document or range formatting (which typically would show up as an annoying
+  toast notification in your code editor) (#120).
 
 # 0.1.1